runtime/vm: extract VM eval loop into vm_eval.go
Pure code motion. vm_eval.go now carries the runtime half of the
package (VM struct, init/SetNowSeed, regTag tags, smallJoinThreshold,
maxCallDepth, New, NewWithIO, Run, RunResult, frame, the dispatch
loop). vm.go keeps the compiler.

vm_eval.go: 1724 lines.
vm.go: 7042 lines (down from 8751 in the parent branch, 9319 on main).

Stacked on #21371.

Refs #21380
